Add build instructions and script to run GEM jekyll.
This builds in to the .build/gh-pages/ directory. However, we do not
yet ensure that this is a checkout of the gh-pages branch of this repo,
nor, even if it were, do we have any method of checking in the build and
pushing it up to GitHub to be served.