Move source for site to src/ subdirectory.
We now generate the gh-pages branch's README.md from our source
here, rather than having to edit it in the gh-pages branch.

From this point on there should be nothing we edit directly in the
gh-pages branch; all releases (i.e., commits to that branch) will be
done by the build system on the master branch.