Add a 'release' Jekyll subcommand (which currently does nothing).
This will be used to push the current build up to a production web
server (assuming that the build is from an unmodified source tree).

We have to do a nasty hack to add a command, because we don't understand
the way you're "suppposed" to do it with gems. Instead we just replace
the standard Jekyll startup script with our own, pulling in our plugin
in the middle of the Jekyll gem initialization.

We also add a _config.yaml option to use an _plugins/ directory at
the top level, rather than under src/, so that any other plugins
we add can go in the same (obvious) place.