Default page layout for all /blog files is now 'post'.
We also add a default author that is basically just a big warning
message to those who forget to add the `author:` key to the YAML
header at the beginning of each blog post. There ought to be a way to
automatically check for the absence of this, but then again, that's why
we want to write a new template system, isn't it?