Shorten title of first Welcome2FPFW post.
Interestingly enough, Jekyll ignores capitalisation in filenames used
as blog post titles and just does its own (usually incorrect)
re-capitalisation for the title, so we specify it properly in the YAML
front matter.

We need to specify the title internally because there's some interesting
auto-caps stuff going on with the filename (ignoring case-sensitivity in
filenames). Hm.