page-head.html: Add a navigation bar.
We also tweak site.css to move the max-width to the body element, so it
applies to the nav bar as well as the content below it.

The "Blog" navigation item just brings up directory listings at the
moment, which is pretty hokey, but will be fixed when we do a proper
top-level blog page.