site.css: center site when in very wide windows.
We still maintian the max-width of 45 ems, and this does not affect
the proper wrapping of everything when the viewport is narrow.

We also change margin to padding since mobile devices are happy to
drop the margin at either side in the hope of getting more text on the
screen, but our site is simple and "pure" enough that it works well
on even the narrowest devices even with a 1-em margin on either side,
rather than having the text be whacked right up against the edges of the
screen (ew).